+++ R/import_query.R
@@ -16,7 +16,6 @@
 #' @param x String containing the path to the Viva Insights query to be
 #'   imported. The input file must be a .csv file, and the file extension must
 #'   be explicitly entered, e.g. `"/files/standard query.csv"`
-#'
 #' @param pid String specifying the unique person or individual identifier
 #'   variable. `import_query` renames this to `PersonId` so that this is
 #'   compatible with other functions in the package. Defaults to `NULL`, where
@@ -24,15 +23,12 @@
 #' @param dateid String specifying the date variable. `import_query` renames
 #'   this to `MetricDate` so that this is compatible with other functions in the
 #'   package. Defaults to `NULL`, where no action is taken.
-#' @param date_format String specifying the date format for converting any
-#' variable that may be a date to a Date variable. Defaults to `"%m/%d/%Y"`.
-#' @param convert_date Logical. Defaults to `TRUE`. When set to `TRUE`, any
-#' variable that matches true with `is_date_format()` gets converted to a Date
-#' variable. When set to `FALSE`, this step is skipped.
+#' @inheritParams prep_query
 #'
 #' @param encoding String to specify encoding to be used within
 #'   `data.table::fread()`. See `data.table::fread()` documentation for more
 #'   information. Defaults to `'UTF-8'`.
+#'   
 #'
 #' @return A `tibble` is returned.
 #'
@@ -46,31 +42,12 @@ import_query <- function(x,
                          convert_date = TRUE,
                          encoding = 'UTF-8') {
 
-
   # import data
   return_data <-
     data.table::fread(x,
                       stringsAsFactors = FALSE,
                       encoding = encoding) %>%
     as.data.frame()
-
-  if(convert_date == TRUE){
-
-    # Columns which are Dates
-    dateCols <- sapply(return_data, function(x) all(is_date_format(x) | any_idate(x)))
-
-    dateCols <- names(return_data)[dateCols == TRUE]
-
-    # Format any date columns
-    return_data <-
-      return_data %>%
-      dplyr::mutate_at(dplyr::vars(dateCols), ~as.Date(., format = date_format))
-
-    if(length(dateCols) >= 1){
-      message("Converted the following Date variables:\n",
-              paste(dateCols, collapse = ", "))
-    }
-  }
 
   # rename specified names
   if(!is.null(pid)){
@@ -80,17 +57,12 @@ import_query <- function(x,
   if(!is.null(dateid)){
     names(return_data)[names(return_data) == dateid] <- "MetricDate"
   }
-
-  # clean names
-  names(return_data) <-
-    names(return_data) %>%
-    gsub(pattern = " ", replacement = "_", x = .) %>% # replace spaces
-    gsub(pattern = "-", replacement = "_", x = .) %>% # replace hyphens
-    gsub(pattern = "[:|,]", replacement = "_", x = .) # replace : and ,
-
-  message("Query has been imported successfully!")
-
-  dplyr::as_tibble(return_data)
+  
+  prep_query(
+    data = return_data,
+    date_format = date_format,
+    convert_date = convert_date
+  )
 }