+++ R/one2one_line.R
@@ -12,7 +12,7 @@
 #' Additional options available to return a summary table.
 #'
 #' @details
-#' Uses the metric `Meeting_hours_with_manager_1_1`.
+#' Uses the metric `Meeting_and_call_hours_with_manager_1_1`.
 #'
 #' @inheritParams create_line
 #' @inherit create_line return
@@ -34,7 +34,7 @@ one2one_line <- function(data,
                          return = "plot"){
   ## Inherit arguments
   create_line(data = data,
-              metric = "Meeting_hours_with_manager_1_1",
+              metric = "Meeting_and_call_hours_with_manager_1_1",
               hrvar = hrvar,
               mingroup = mingroup,
               return = return)