+++ R/one2one_sum.R
@@ -32,13 +32,13 @@ one2one_sum <- function(data,
 
   cleaned_data <-
     data %>%
-    mutate(`Scheduled 1:1 meeting minutes with manager` = Meeting_hours_with_manager_1_1 * 60)
+    mutate(`Scheduled 1:1 meeting and call minutes with manager` = Meeting_and_call_hours_with_manager_1_1 * 60)
 
 
   create_bar(data = cleaned_data,
              hrvar = hrvar,
              mingroup = mingroup,
-             metric = "Scheduled 1:1 meeting minutes with manager",
+             metric = "Scheduled 1:1 meeting and call minutes with manager",
              return = return,
              bar_colour = "darkblue")