+++ R/us_to_space.R
@@ -10,7 +10,7 @@
 #' @family Support
 #'
 #' @examples
-#' us_to_space("Meeting_hours_with_manager_1_on_1")
+#' us_to_space("Meeting_and_call_hours_with_manager_1_on_1")
 #'
 #' @export
 us_to_space <- function(x){